+++ pkg/plugins/trivy/image.go
@@ -3,6 +3,8 @@ package trivy
 import (
 	"fmt"
 	"net/url"
+	"regexp"
+	"strings"
 
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/utils/pointer"
@@ -90,7 +92,7 @@ func GetPodSpecForStandaloneMode(ctx trivyoperator.PluginContext, config Config,
 		Image:                    trivyImageRef,
 		ImagePullPolicy:          corev1.PullPolicy(config.GetImagePullPolicy()),
 		TerminationMessagePolicy: corev1.TerminationMessageFallbackToLogsOnError,
-		Env:                      p.initContainerEnvVar(trivyConfigName, config),
+		Env:                      initContainerEnvVar(trivyConfigName, config),
 		Command: []string{
 			"trivy",
 		},
@@ -227,12 +229,12 @@ func GetPodSpecForStandaloneMode(ctx trivyoperator.PluginContext, config Config,
 
 		}
 
-		env, err = p.appendTrivyInsecureEnv(config, c.Image, env)
+		env, err = appendTrivyInsecureEnv(config, c.Image, env)
 		if err != nil {
 			return corev1.PodSpec{}, nil, err
 		}
 
-		env, err = p.appendTrivyNonSSLEnv(config, c.Image, env)
+		env, err = appendTrivyNonSSLEnv(config, c.Image, env)
 		if err != nil {
 			return corev1.PodSpec{}, nil, err
 		}
@@ -247,7 +249,7 @@ func GetPodSpecForStandaloneMode(ctx trivyoperator.PluginContext, config Config,
 			return corev1.PodSpec{}, nil, err
 		}
 		resultFileName := getUniqueScanResultFileName(c.Name)
-		cmd, args := p.getCommandAndArgs(ctx, Standalone, imageRef.String(), "", resultFileName)
+		cmd, args := getCommandAndArgs(ctx, Standalone, imageRef.String(), "", resultFileName)
 		containers = append(containers, corev1.Container{
 			Name:                     c.Name,
 			Image:                    trivyImageRef,
@@ -426,12 +428,12 @@ func GetPodSpecForClientServerMode(ctx trivyoperator.PluginContext, config Confi
 			}
 		}
 
-		env, err = p.appendTrivyInsecureEnv(config, container.Image, env)
+		env, err = appendTrivyInsecureEnv(config, container.Image, env)
 		if err != nil {
 			return corev1.PodSpec{}, nil, err
 		}
 
-		env, err = p.appendTrivyNonSSLEnv(config, container.Image, env)
+		env, err = appendTrivyNonSSLEnv(config, container.Image, env)
 		if err != nil {
 			return corev1.PodSpec{}, nil, err
 		}
@@ -457,7 +459,7 @@ func GetPodSpecForClientServerMode(ctx trivyoperator.PluginContext, config Confi
 			return corev1.PodSpec{}, nil, err
 		}
 		resultFileName := getUniqueScanResultFileName(container.Name)
-		cmd, args := p.getCommandAndArgs(ctx, ClientServer, imageRef.String(), encodedTrivyServerURL.String(), resultFileName)
+		cmd, args := getCommandAndArgs(ctx, ClientServer, imageRef.String(), encodedTrivyServerURL.String(), resultFileName)
 		containers = append(containers, corev1.Container{
 			Name:                     container.Name,
 			Image:                    trivyImageRef,
@@ -480,4 +482,224 @@ func GetPodSpecForClientServerMode(ctx trivyoperator.PluginContext, config Confi
 		Containers:                   containers,
 		Volumes:                      volumes,
 	}, secrets, nil
+}
+
+func initContainerEnvVar(trivyConfigName string, config Config) []corev1.EnvVar {
+	envs := []corev1.EnvVar{
+		constructEnvVarSourceFromConfigMap("HTTP_PROXY", trivyConfigName, keyTrivyHTTPProxy),
+		constructEnvVarSourceFromConfigMap("HTTPS_PROXY", trivyConfigName, keyTrivyHTTPSProxy),
+		constructEnvVarSourceFromConfigMap("NO_PROXY", trivyConfigName, keyTrivyNoProxy),
+		constructEnvVarSourceFromSecret("GITHUB_TOKEN", trivyConfigName, keyTrivyGitHubToken),
+	}
+
+	if config.GetDBRepositoryInsecure() {
+		envs = append(envs, corev1.EnvVar{
+			Name:  "TRIVY_INSECURE",
+			Value: "true",
+		})
+	}
+	return envs
+}
+
+func getCommandAndArgs(ctx trivyoperator.PluginContext, mode Mode, imageRef string, trivyServerURL string, resultFileName string) ([]string, []string) {
+	command := []string{
+		"trivy",
+	}
+	trivyConfig := ctx.GetTrivyOperatorConfig()
+	compressLogs := trivyConfig.CompressLogs()
+	c, err := getConfig(ctx)
+	if err != nil {
+		return []string{}, []string{}
+	}
+	slow := Slow(c)
+	skipJavaDBUpdate := SkipJavaDBUpdate(c)
+	vulnTypeArgs := vulnTypeFilter(ctx)
+	scanners := Scanners(c)
+	var vulnTypeFlag string
+	if len(vulnTypeArgs) == 2 {
+		vulnTypeFlag = fmt.Sprintf("%s %s ", vulnTypeArgs[0], vulnTypeArgs[1])
+	}
+	imcs := imageConfigSecretScanner(trivyConfig)
+	var imageconfigSecretScannerFlag string
+	if len(imcs) == 2 {
+		imageconfigSecretScannerFlag = fmt.Sprintf("%s %s ", imcs[0], imcs[1])
+	}
+	var skipUpdate string
+	if mode == ClientServer {
+		if c.GetClientServerSkipUpdate() {
+			skipUpdate = SkipDBUpdate(c)
+		}
+		if !compressLogs {
+			args := []string{
+				"--cache-dir",
+				"/tmp/trivy/.cache",
+				"--quiet",
+				"image",
+				scanners,
+				getSecurityChecks(ctx),
+				"--format",
+				"json",
+				"--server",
+				trivyServerURL,
+				imageRef,
+			}
+			if len(slow) > 0 {
+				args = append(args, slow)
+			}
+			if len(vulnTypeArgs) > 0 {
+				args = append(args, vulnTypeArgs...)
+			}
+			if len(imcs) > 0 {
+				args = append(args, imcs...)
+			}
+			pkgList := getPkgList(ctx)
+			if len(pkgList) > 0 {
+				args = append(args, pkgList)
+			}
+			if len(skipUpdate) > 0 {
+				args = append(args, skipUpdate)
+			}
+			if len(skipJavaDBUpdate) > 0 {
+				args = append(args, skipJavaDBUpdate)
+			}
+
+			return command, args
+		}
+		return []string{"/bin/sh"}, []string{"-c", fmt.Sprintf(`trivy image %s '%s' %s %s %s %s %s %s --cache-dir /tmp/trivy/.cache --quiet %s --format json --server '%s' > /tmp/scan/%s &&  bzip2 -c /tmp/scan/%s | base64`, slow, imageRef, scanners, getSecurityChecks(ctx), imageconfigSecretScannerFlag, vulnTypeFlag, skipUpdate, skipJavaDBUpdate, getPkgList(ctx), trivyServerURL, resultFileName, resultFileName)}
+	}
+	skipUpdate = SkipDBUpdate(c)
+	if !compressLogs {
+		args := []string{
+			"--cache-dir",
+			"/tmp/trivy/.cache",
+			"--quiet",
+			"image",
+			scanners,
+			getSecurityChecks(ctx),
+			"--format",
+			"json",
+			imageRef,
+		}
+		if len(slow) > 0 {
+			args = append(args, slow)
+		}
+		if len(vulnTypeArgs) > 0 {
+			args = append(args, vulnTypeArgs...)
+		}
+		if len(imcs) > 0 {
+			args = append(args, imcs...)
+		}
+		pkgList := getPkgList(ctx)
+		if len(pkgList) > 0 {
+			args = append(args, pkgList)
+		}
+		if len(skipUpdate) > 0 {
+			args = append(args, skipUpdate)
+		}
+		if len(skipJavaDBUpdate) > 0 {
+			args = append(args, skipJavaDBUpdate)
+		}
+		return command, args
+	}
+	return []string{"/bin/sh"}, []string{"-c", fmt.Sprintf(`trivy image %s '%s' %s %s %s %s %s %s --cache-dir /tmp/trivy/.cache --quiet %s --format json > /tmp/scan/%s &&  bzip2 -c /tmp/scan/%s | base64`, slow, imageRef, scanners, getSecurityChecks(ctx), imageconfigSecretScannerFlag, vulnTypeFlag, skipUpdate, skipJavaDBUpdate, getPkgList(ctx), resultFileName, resultFileName)}
+}
+
+func getConfig(ctx trivyoperator.PluginContext) (Config, error) {
+	pluginConfig, err := ctx.GetConfig()
+	if err != nil {
+		return Config{}, err
+	}
+	return Config{PluginConfig: pluginConfig}, nil
+}
+
+func vulnTypeFilter(ctx trivyoperator.PluginContext) []string {
+	config, err := getConfig(ctx)
+	if err != nil {
+		return []string{}
+	}
+	vulnType := config.GetVulnType()
+	if len(vulnType) == 0 {
+		return []string{}
+	}
+	return []string{"--vuln-type", vulnType}
+}
+
+func appendTrivyNonSSLEnv(config Config, image string, env []corev1.EnvVar) ([]corev1.EnvVar, error) {
+	ref, err := containerimage.ParseReference(image)
+	if err != nil {
+		return nil, err
+	}
+
+	nonSSLRegistries := config.GetNonSSLRegistries()
+	if nonSSLRegistries[ref.Context().RegistryStr()] {
+		env = append(env, corev1.EnvVar{
+			Name:  "TRIVY_NON_SSL",
+			Value: "true",
+		})
+	}
+	return env, nil
+}
+
+func createEnvandVolumeForGcr(env *[]corev1.EnvVar, volumeMounts *[]corev1.VolumeMount, volumes *[]corev1.Volume, registryPasswordKey *string, secretName *string) {
+	*env = append(*env, corev1.EnvVar{
+		Name:  "TRIVY_USERNAME",
+		Value: "",
+	})
+	*env = append(*env, corev1.EnvVar{
+		Name:  "GOOGLE_APPLICATION_CREDENTIALS",
+		Value: "/cred/credential.json",
+	})
+	googlecredMount := corev1.VolumeMount{
+		Name:      "gcrvol",
+		MountPath: "/cred",
+		ReadOnly:  true,
+	}
+	googlecredVolume := corev1.Volume{
+		Name: "gcrvol",
+		VolumeSource: corev1.VolumeSource{
+			Secret: &corev1.SecretVolumeSource{
+				SecretName: *secretName,
+				Items: []corev1.KeyToPath{
+					{
+						Key:  *registryPasswordKey,
+						Path: "credential.json",
+					},
+				},
+			},
+		},
+	}
+	*volumes = append(*volumes, googlecredVolume)
+	*volumeMounts = append(*volumeMounts, googlecredMount)
+}
+
+func checkGcpCrOrPivateRegistry(imageUrl string) bool {
+	imageRegex := regexp.MustCompile(GCPCR_Inage_Regex)
+	return imageRegex.MatchString(imageUrl)
+}
+
+func getUniqueScanResultFileName(name string) string {
+	return fmt.Sprintf("result_%s.json", name)
+}
+
+func CheckAwsEcrPrivateRegistry(ImageUrl string) string {
+	if len(regexp.MustCompile(AWSECR_Image_Regex).FindAllStringSubmatch(ImageUrl, -1)) != 0 {
+		return regexp.MustCompile(AWSECR_Image_Regex).FindAllStringSubmatch(ImageUrl, -1)[0][1]
+	}
+	return ""
+}
+
+func GetMirroredImage(image string, mirrors map[string]string) (string, error) {
+	ref, err := containerimage.ParseReference(image)
+	if err != nil {
+		return "", err
+	}
+	mirroredImage := ref.Name()
+	for k, v := range mirrors {
+		if strings.HasPrefix(mirroredImage, k) {
+			mirroredImage = strings.Replace(mirroredImage, k, v, 1)
+			return mirroredImage, nil
+		}
+	}
+	// If nothing is mirrored, we can simply use the input image.
+	return image, nil
 }
